backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.5.3
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
psycopg[binary]
uvicorn
a2wsgi
python-dotenv==1.0.0
bcrypt==4.1.2
pypdf==5.4.0
